Fix errors in foreign key binding

+++ database/migrations/2023_03_30_232918_create_todos_table.php
@@ -15,7 +15,7 @@ class CreateTodosTable extends Migration
     {
         Schema::create('todos', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('task_id');
+            $table->unsignedBigInteger('task_id');
             $table->enum('status', ['0', '1'])->default('0')->comment('O means Not Started, 1 means Completed');
             $table->string('name');
             $table->timestamps();
